requirements.txt
alabaster==0.7.12
alembic 
aniso8601==8.0.0
astroid 
attrs
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach
blinker==1.4
brotlipy==0.7.0
certifi==2020.6.20
cffi
chardet==3.0.4
click==7.1.2
cloudpickle
colorama==0.4.3
cryptography
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
docutils==0.16
entrypoints==0.3
Flask==1.1.2
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-RESTful==0.3.8
Flask-SQLAlchemy 
Flask-WTF==0.14.2
gunicorn==20.0.4
httplib2==0.11.3
idna 
imagesize==1.2.0
importlib-metadata
ipykernel
ipython
ipython-genutils==0.2.0
isort
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.11.2
Js2Py==0.70
jsonschema==3.2.0
jupyter-client
jupyter-core==4.6.3
keyring
lazy==1.3
lazy-object-proxy==1.4.3
Mako 
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
numpydoc
oauth2==1.9.0.post1
oauthlib==2.0.7
packaging==20.4
pandocfilters==1.4.2
parso
pickleshare==0.7.5
prompt-toolkit 
psutil
pycodestyle==2.6.0
pycparser
pyflakes==2.2.0
Pygments 
pyjsparser==2.7.1
PyJWT==1.4.2
pylint
pyOpenSSL 
pyparsing==2.4.7
PyPrind==2.11.2
pyrsistent
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
pyzmq==19.0.2
QtAwesome==0.7.2
qtconsole
QtPy==1.9.0
requests 
requests-oauthlib==0.8.0
rope==0.17.0
sip==4.19.13
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx 
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy 
SQLAlchemy-Utils==0.33.2
testpath==0.4.4
toml 
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.1
urllib3 
URLObject==2.4.3
wcwidth 
webencodings==0.5.1
Werkzeug==0.16.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
WTForms==2.1
zipp==3.1.0
